requirements.txt
torch==2.2.2
torchvision==0.17.2
numpy==1.24.3
scikit-learn==1.3.2
scikit-image==0.21.0
matplotlib==3.7.5
pandas==2.0.3
openpyxl==3.1.5
joblib==1.4.2
tqdm==4.66.5
python-dateutil==2.9.0.post0
pillow==10.2.0
seaborn==0.13.2
requests==2.31.0
